fixing some pugs

+++ controllers/post.js
@@ -23,53 +23,69 @@ const uploadMedia = (fileBuffer, resourceType) => {
 
 const create = async (req, res) => {
   try {
-    let postData = {}
+    const ownerId = req.user._id
 
-    postData.owner = req.session.user._id
-    postData.media = {}
+    let media = {}
 
     if (req.body.text) {
-      postData.media.type = "text"
-      postData.media.text = req.body.text
+      media.type = "text"
+      media.text = req.body.text
     } else if (req.file) {
-      const mediaType = req.file.mimetype.startsWith("video/")
-        ? "video"
-        : "image"
+      let mediaType = "image"
+      if (req.file.mimetype.startsWith("video/")) {
+        mediaType = "video"
+      }
 
       const result = await uploadMedia(req.file.buffer, mediaType)
 
-      postData.media.type = mediaType
-      postData.media.url = result.secure_url
-      postData.media.publicId = result.public_id
+      media.type = mediaType
+      media.url = result.secure_url
+      media.publicId = result.public_id
     }
 
-    const post = await Post.create(postData)
+    const newPostData = {
+      owner: ownerId,
+      media: media,
+    }
+
+    const post = await (await Post.create(newPostData)).populate('owner')
+
     res.status(201).json(post)
-  } catch (error) {
-    res.status(500).json({ error: error.message })
+  } catch (err) {
+    res.status(500).json({ err: err.message })
   }
 }
+
 const index = async (req, res) => {
-    try {
-        
-        const posts = await Post.find().populate("owner")
-        res.status(200).json(posts)
-    } catch (error) {
-         res.status(500).json({ err: err.message })
-    }
+  try {
+    const posts = await Post.find().populate("owner")
+    res.status(200).json(posts)
+  } catch (error) {
+    res.status(500).json({ err: error.message })
+  }
 }
 
 const show = async (req, res) => {
   try {
-    const post = await Post.findById(req.params.hootId).populate('author')
+    const post = await Post.findById(req.params.postId).populate('owner')
+
+    if (!post) {
+      return res.status(404).json({ err: 'Post not found' })
+    }
+
     res.status(200).json(post)
   } catch (err) {
     res.status(500).json({ err: err.message })
   }
 }
+
 const update = async (req, res) => {
   try {
     const post = await Post.findById(req.params.postId)
+
+    if (!post) {
+      return res.status(404).json({ err: 'Post not found' })
+    }
 
     if (!post.owner.equals(req.user._id)) {
       return res.status(403).send("You're not allowed to do that!")
@@ -80,8 +96,9 @@ const update = async (req, res) => {
       req.body,
       { new: true }
     )
+    updatedPost._doc.owner = req.user
 
-    res.status(202).json(updatedPost)
+    res.status(200).json(updatedPost)
   } catch (err) {
     res.status(500).json({ err: err.message })
   }
@@ -91,20 +108,25 @@ const deletePost = async (req, res) => {
   try {
     const post = await Post.findById(req.params.postId)
 
+    if (!post) {
+      return res.status(404).json({ err: 'Post not found' })
+    }
+
     if (!post.owner.equals(req.user._id)) {
       return res.status(403).send("You're not allowed to do that!")
     }
 
-    const deletePost = await Post.findByIdAndDelete(req.params.postId)
-    res.status(200).json(deletePost)
+    const deletedPost = await Post.findByIdAndDelete(req.params.postId)
+    res.status(200).json(deletedPost)
   } catch (err) {
     res.status(500).json({ err: err.message })
   }
 }
+
 module.exports = {
-    create,
-    index,
-    show,
-    update,
-    deletePost
+  create,
+  index,
+  show,
+  update,
+  deletePost
 }
